dashboard/requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
google-cloud-bigquery>=3.17.0
google-cloud-bigquery-storage>=2.24.0
pandas>=2.0.0
pyarrow>=15.0.0
db-dtypes>=1.2.0
statsmodels>=0.14.0